test/api/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
mysql-connector-python==9.0.0